% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{calc.q}
\alias{calc.q}
\title{Calculate mu, based on a certain model}
\usage{
calc.q(df = NULL, calc.signal = NULL, model = "gompertz")
}
\arguments{
\item{df}{}

\item{calc.signal}{}

\item{model}{}
}
\description{
Calculate mu, based on a certain model
}
