% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{calc.ratio}
\alias{calc.ratio}
\title{calcuate ratio q}
\usage{
calc.ratio(df = NULL, x.signal = NULL, y.signal = NULL)
}
\arguments{
\item{df}{}

\item{x.signal}{}

\item{y.signal}{}
}
\description{
calcuate ratio q
}
