% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{extract.ratio.data}
\alias{extract.ratio.data}
\title{extract ratio data}
\usage{
\method{extract}{ratio.data}(df, col)
}
\arguments{
\item{df}{}

\item{col}{}
}
\description{
extract ratio data
}
\author{
Pieter Coussement, \email{coussementpieter@gmail.com}
}
