% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{q.calc.individual}
\alias{q.calc.individual}
\title{Calculate mu, based on a certain model}
\usage{
q.calc.individual(df = NULL, calc.signal = NULL, model = "gompertz")
}
\arguments{
\item{df}{}

\item{calc.signal}{}

\item{model}{}
}
\description{
Calculate mu, based on a certain model
}
