% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{q.extract.data}
\alias{q.extract.data}
\title{Extract calculated growth parameters into a dataframe}
\usage{
q.extract.data(df, col)
}
\arguments{
\item{df}{}

\item{col}{}
}
\description{
Extract calculated growth parameters into a dataframe
}
