% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{q.extract.q.data.individual}
\alias{q.extract.q.data.individual}
\title{Extract calculated growth parameters into a dataframe}
\usage{
q.extract.q.data.individual(grofit.model)
}
\arguments{
\item{df}{}

\item{col}{}
}
\description{
Extract calculated growth parameters into a dataframe
}
\author{
Pieter Coussement, \email{coussementpieter@gmail.com}
}
