% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{q.plot.individual}
\alias{q.plot.individual}
\title{plot the q fit}
\usage{
q.plot.individual(fit = NULL, calc.signal = NULL, name = NULL)
}
\arguments{
\item{fit}{}

\item{calc.signal}{}

\item{name}{}
}
\description{
plot the q fit
}
