% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ratio.calc.individual}
\alias{ratio.calc.individual}
\title{calcuate ratio q}
\usage{
ratio.calc.individual(df, x.signal = NULL, y.signal = NULL)
}
\arguments{
\item{df}{}

\item{x.signal}{}

\item{y.signal}{}
}
\description{
calcuate ratio q
}
