% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ratio.extract.data}
\alias{ratio.extract.data}
\title{extract ratio data}
\usage{
ratio.extract.data(df, col)
}
\arguments{
\item{df}{}

\item{col}{}
}
\description{
extract ratio data
}
\author{
Pieter Coussement, \email{coussementpieter@gmail.com}
}
