% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{calc.exp.time}
\alias{calc.exp.time}
\title{Title}
\usage{
calc.exp.time(df = NULL, input_time = "time", input_value = "value",
  treshold = 0.3, lambda = 1e-05)
}
\arguments{
\item{df}{}

\item{input_time}{}

\item{treshold}{}

\item{signal}{}
}
\description{
Title
}
