% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{calc.exp.time}
\alias{calc.exp.time}
\title{Title}
\usage{
calc.exp.time(df = NULL, signal = NULL, input_time = "time_hour",
  treshold = 0.3)
}
\arguments{
\item{df}{}

\item{signal}{}

\item{input_time}{}

\item{treshold}{}
}
\description{
Title
}
