% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{q.calc.individual}
\alias{q.calc.individual}
\title{Calculate mu, based on a certain model}
\usage{
q.calc.individual(df, model = "gompertz")
}
\arguments{
\item{df}{(dataframe). dataframe containing all necessary data for the calculation}

\item{model}{(string). model to is used to fit. (default: "gompertz")}
}
\value{
grofit object
}
\description{
Calculate mu, based on a certain model
}
\author{
Pieter Coussement, \email{coussementpieter@gmail.com}
}
