% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyse.R
\name{ratio.calc.individual}
\alias{ratio.calc.individual}
\title{calcuate ratio q}
\usage{
ratio.calc.individual(df, x.signal = NULL, y.signal = NULL)
}
\arguments{
\item{df}{(datafrane). Dataframe containing the data}

\item{x.signal}{(string). column name to be used in the x.axis}

\item{y.signal}{(string). column name to be used in the y.axis}
}
\value{
fitobject
}
\description{
calcuate ratio q
}
\author{
Pieter Coussement, \email{coussementpieter@gmail.com}
}
